requirements.txt
absl-py==1.0.0
av==9.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.6.0
cycler==0.11.0
Flask==2.0.3
Flask-WTF==1.0.0
fonttools==4.31.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.19.0
gym-notices==0.0.6
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
jsonpickle==2.1.0
kiwisolver==1.4.0
lark==1.1.2
lark-parser==0.12.0
llvmlite==0.38.0
ltlf2dfa==1.0.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mpmath==1.2.1
networkx==2.6.3
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
palettable==3.3.0
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sympy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tifffile==2021.11.2
torch==1.11.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
